[CHANGE] Switched to wrapper types in DTOs

+++ mobile/app/src/main/java/com/example/ubre/ui/dtos/CancellationDto.java
@@ -3,21 +3,21 @@ package com.example.ubre.ui.dtos;
 // DTO for ride cancellation requests
 
 public class CancellationDto {
-    private int rideId;
+    private Long rideId;
     private String reason;
 
     public CancellationDto() {
     }
 
-    public CancellationDto(int rideId, String reason) {
+    public CancellationDto(Long rideId, String reason) {
         this.rideId = rideId;
         this.reason = reason;
     }
 
-    public int getRideId() {
+    public Long getRideId() {
         return rideId;
     }
-    public void setRideId(int rideId) {
+    public void setRideId(Long rideId) {
         this.rideId = rideId;
     }
     public String getReason() {
