+++ mobile/app/src/main/java/com/example/ubre/ui/dtos/RideCardDto.java
@@ -2,22 +2,21 @@ package com.example.ubre.ui.dtos;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Date;
 
 // Ride history, active rides, my favourites
 
 public class RideCardDto {
-    private int id;
+    private Long id;
     private LocalDateTime startTime;
     public ArrayList<WaypointDto> waypoints;
 
-    public RideCardDto(int id, LocalDateTime startTime, ArrayList<WaypointDto> waypoints) {
+    public RideCardDto(Long id, LocalDateTime startTime, ArrayList<WaypointDto> waypoints) {
         this.id = id;
         this.startTime = startTime;
         this.waypoints = waypoints;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
     public LocalDateTime getStartTime() {
@@ -27,7 +26,7 @@ public class RideCardDto {
         return waypoints;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public void setStartTime(LocalDateTime startTime) {
